+++ docs/sections/Docs-Slate/index.jsx
@@ -50,6 +50,16 @@ export default class Slate extends React.Component {
             <small>Centered slate with 50% width</small>
           </div>
         </div>
+
+        <div className="doc-bottom-space">
+          <h3 className="doc-heading">Slate With Icons</h3>
+          <p>Icons are automatically given a margin of `$spacer-large`.</p>
+          <div className="slate col-md-6 centered">
+            <i className="icon icon-cog -x8" />
+            <h2>Centered Half Slate</h2>
+            <small>Centered slate with 50% width</small>
+          </div>
+        </div>
       </div>
     );
   }